Refractor Time

+++ app/Http/Controllers/ManageUsersController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\User;
+use Illuminate\Support\Facades\Auth;
 use Hash;
 use Illuminate\Http\Request;
 
@@ -42,9 +43,9 @@ class ManageUsersController extends Controller
     {
         // Validate
         $this->validate($request, [
-            'name' => 'required|min:3|max:255',
-            'email' => 'required|email|unique:users', // Two users can't have the same email
-            'password' => 'required|min:8'
+            'name'                  => 'required|min:3|max:255',
+            'email'                 => 'required|email|unique:users', // Two users can't have the same email
+            'password'              => 'required|confirmed|min:8',
         ]);
 
         $store = $request->all();
@@ -92,8 +93,9 @@ class ManageUsersController extends Controller
     {
         // Validate values
         $this->validate($request, [
-            'name' => 'required|min:3|max:255',
-            'email' => 'required|email|unique:users,email,'.$user->id // Two users can't have the same email
+            'name'                  => 'required|min:3|max:255',
+            'email'                 => 'required|email|unique:users,email,'.$user->id, // Two users can't have the same email
+            'password'              => 'required|confirmed|min:8',
         ]);
 
         $store = $request->all();
@@ -118,7 +120,15 @@ class ManageUsersController extends Controller
      */
     public function destroy(User $user)
     {
-        $user->delete();
-        return redirect()->route('users.index')->with('sucess', 'USER HAS BEEN KILLED >:D');
+
+        // $user->delete();
+        // return redirect()->route('users.index')->with('sucess', 'USER HAS BEEN KILLED >:D');
+
+        if(Auth::id() != $user->id) {
+            $user->delete();
+            return redirect()->route('users.index')->with('sucess', 'USER HAS BEEN KILLED >:D');
+        } else {
+            return redirect()->route('users.index')->with('fail', 'USER CAN\'T KILL THEMSELFS');
+        }
     }
 }
